reorganizing and removing duplicate code
