+++ next-hardhat/.prettierignore
