+++ next-hardhat/pages/api/listDirectory.ts
