+++ next-hardhat/pages/api/readCircuitFile.ts
