+++ next-hardhat/utils/addresses.json
@@ -1,0 +1,1 @@
+{"chainId":31337,"verifier":"0xB7f8BC63BbcaD18155201308C8f3540b07f84F5e"}
