+++ next-hardhat/wasm.d.ts
