--- utils/addresses.json
@@ -1,1 +1,0 @@
-{"chainId":80001,"verifier":"0xb6f9DB74DBCa7AF4DC1B1B5A3576F5DCCFA0355E"}
